storage/requirements.txt
a2wsgi==1.8.0
anyio==4.0.0
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
connexion==3.0.0
Flask==3.0.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kazoo==2.5.0
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
pykafka==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.31.1
swagger_ui_bundle==1.1.0
tabulate==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.24.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
